[Optimization-2579][core] Optimizate JobManager
