+++ dinky-core/src/main/java/org/dinky/job/JobManager.java
@@ -19,10 +19,6 @@
 
 package org.dinky.job;
 
-import static org.dinky.function.util.UDFUtil.GATEWAY_TYPE_MAP;
-import static org.dinky.function.util.UDFUtil.SESSION;
-import static org.dinky.function.util.UDFUtil.YARN;
-
 import org.dinky.api.FlinkAPI;
 import org.dinky.assertion.Asserts;
 import org.dinky.constant.FlinkSQLConstant;
@@ -31,8 +27,6 @@ import org.dinky.context.FlinkUdfPathContextHolder;
 import org.dinky.context.RowLevelPermissionsContext;
 import org.dinky.data.annotations.ProcessStep;
 import org.dinky.data.enums.ProcessStepType;
-import org.dinky.data.exception.DinkyException;
-import org.dinky.data.model.FlinkUdfManifest;
 import org.dinky.data.model.SystemConfiguration;
 import org.dinky.data.result.ErrorResult;
 import org.dinky.data.result.ExplainResult;
@@ -43,11 +37,9 @@ import org.dinky.data.result.ResultPool;
 import org.dinky.data.result.SelectResult;
 import org.dinky.executor.Executor;
 import org.dinky.executor.ExecutorConfig;
+import org.dinky.executor.ExecutorContext;
 import org.dinky.executor.ExecutorFactory;
 import org.dinky.explainer.Explainer;
-import org.dinky.function.constant.PathConstant;
-import org.dinky.function.data.model.UDF;
-import org.dinky.function.util.UDFUtil;
 import org.dinky.gateway.Gateway;
 import org.dinky.gateway.config.FlinkConfig;
 import org.dinky.gateway.config.GatewayConfig;
@@ -57,57 +49,38 @@ import org.dinky.gateway.enums.SavePointType;
 import org.dinky.gateway.result.GatewayResult;
 import org.dinky.gateway.result.SavePointResult;
 import org.dinky.gateway.result.TestResult;
-import org.dinky.interceptor.FlinkInterceptor;
-import org.dinky.interceptor.FlinkInterceptorResult;
+import org.dinky.job.builder.JobDDLBuilder;
+import org.dinky.job.builder.JobExecuteBuilder;
+import org.dinky.job.builder.JobJarStreamGraphBuilder;
+import org.dinky.job.builder.JobTransBuilder;
+import org.dinky.job.builder.JobUDFBuilder;
 import org.dinky.parser.SqlType;
 import org.dinky.trans.Operations;
-import org.dinky.trans.dml.ExecuteJarOperation;
-import org.dinky.trans.parse.AddJarSqlParseStrategy;
-import org.dinky.trans.parse.ExecuteJarParseStrategy;
 import org.dinky.utils.DinkyClassLoaderUtil;
 import org.dinky.utils.JsonUtils;
 import org.dinky.utils.LogUtil;
 import org.dinky.utils.SqlUtil;
 import org.dinky.utils.URLUtils;
 
-import org.apache.flink.configuration.Configuration;
 import org.apache.flink.configuration.CoreOptions;
 import org.apache.flink.configuration.DeploymentOptions;
 import org.apache.flink.configuration.PipelineOptions;
-import org.apache.flink.core.execution.JobClient;
 import org.apache.flink.runtime.jobgraph.JobGraph;
 import org.apache.flink.runtime.jobgraph.SavepointConfigOptions;
 import org.apache.flink.runtime.jobgraph.SavepointRestoreSettings;
 import org.apache.flink.runtime.jobgraph.jsonplan.JsonPlanGenerator;
 import org.apache.flink.streaming.api.environment.ExecutionCheckpointingOptions;
-import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.streaming.api.graph.StreamGraph;
 import org.apache.flink.table.api.TableResult;
 import org.apache.flink.yarn.configuration.YarnConfigOptions;
 
-import java.io.File;
-import java.lang.reflect.Field;
-import java.net.URL;
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collections;
-import java.util.List;
 import java.util.Map;
-import java.util.Set;
-import java.util.stream.Collectors;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-import cn.hutool.core.collection.CollUtil;
-import cn.hutool.core.io.FileUtil;
-import cn.hutool.core.lang.Assert;
 import cn.hutool.core.text.StrFormatter;
-import cn.hutool.core.util.ArrayUtil;
-import cn.hutool.core.util.RandomUtil;
-import cn.hutool.core.util.StrUtil;
-import cn.hutool.core.util.URLUtil;
-import cn.hutool.json.JSONUtil;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -123,7 +96,34 @@ public class JobManager {
     private String sqlSeparator = FlinkSQLConstant.SEPARATOR;
     private GatewayType runMode = GatewayType.LOCAL;
 
+    private JobParam jobParam = null;
+    private String currentSql = "";
+
     public JobManager() {}
+
+    public JobParam getJobParam() {
+        return jobParam;
+    }
+
+    public void setJobParam(JobParam jobParam) {
+        this.jobParam = jobParam;
+    }
+
+    public JobConfig getConfig() {
+        return config;
+    }
+
+    public GatewayType getRunMode() {
+        return runMode;
+    }
+
+    public void setCurrentSql(String currentSql) {
+        this.currentSql = currentSql;
+    }
+
+    public Executor getExecutor() {
+        return executor;
+    }
 
     public void setPlanMode(boolean planMode) {
         isPlanMode = planMode;
@@ -137,27 +137,19 @@ public class JobManager {
         return useStatementSet;
     }
 
-    public void setUseStatementSet(boolean useStatementSet) {
-        this.useStatementSet = useStatementSet;
-    }
-
     public boolean isUseRestAPI() {
         return useRestAPI;
-    }
-
-    public void setUseRestAPI(boolean useRestAPI) {
-        this.useRestAPI = useRestAPI;
     }
 
     public String getSqlSeparator() {
         return sqlSeparator;
     }
 
-    public void setSqlSeparator(String sqlSeparator) {
-        this.sqlSeparator = sqlSeparator;
+    public boolean isUseGateway() {
+        return useGateway;
     }
 
-    public JobManager(JobConfig config) {
+    private JobManager(JobConfig config) {
         this.config = config;
     }
 
@@ -175,7 +167,7 @@ public class JobManager {
         return manager;
     }
 
-    public boolean init() {
+    public void init() {
         if (!isPlanMode) {
             runMode = GatewayType.get(config.getType());
             useGateway = GatewayType.isDeployCluster(config.getType());
@@ -186,110 +178,7 @@ public class JobManager {
         sqlSeparator = SystemConfiguration.getInstances().getSqlSeparator();
         executorConfig = config.getExecutorSetting();
         executor = ExecutorFactory.buildExecutor(executorConfig);
-        return false;
-    }
-
-    private void addConfigurationClsAndJars(List<URL> jarList, List<URL> classpaths) throws Exception {
-        Map<String, Object> temp = getFlinkConfigurationMap();
-        temp.put(
-                PipelineOptions.CLASSPATHS.key(),
-                classpaths.stream().map(URL::toString).collect(Collectors.toList()));
-        temp.put(PipelineOptions.JARS.key(), jarList.stream().map(URL::toString).collect(Collectors.toList()));
-    }
-
-    private Map<String, Object> getFlinkConfigurationMap() {
-        Field configuration = null;
-        try {
-            configuration = StreamExecutionEnvironment.class.getDeclaredField("configuration");
-            configuration.setAccessible(true);
-            Configuration o = (Configuration) configuration.get(executor.getStreamExecutionEnvironment());
-            Field confData = Configuration.class.getDeclaredField("confData");
-            confData.setAccessible(true);
-            Map<String, Object> temp = (Map<String, Object>) confData.get(o);
-            return temp;
-        } catch (Exception e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    public void initUDF(List<UDF> udfList) {
-        if (Asserts.isNotNullCollection(udfList)) {
-            initUDF(udfList, runMode, config.getTaskId());
-        }
-    }
-
-    public void initUDF(List<UDF> udfList, GatewayType runMode, Integer taskId) {
-        if (taskId == null) {
-            taskId = -RandomUtil.randomInt(0, 1000);
-        }
-        // TODO 改为ProcessStep注释
-
-        // 这里要分开
-        // 1. 得到jar包路径，注入remote环境
-        Set<File> jarFiles = FlinkUdfPathContextHolder.getUdfFile();
-
-        Set<File> otherPluginsFiles = FlinkUdfPathContextHolder.getOtherPluginsFiles();
-        jarFiles.addAll(otherPluginsFiles);
-
-        List<File> udfJars = Arrays.stream(UDFUtil.initJavaUDF(udfList, runMode, taskId))
-                .map(File::new)
-                .collect(Collectors.toList());
-        jarFiles.addAll(udfJars);
-
-        String[] jarPaths = CollUtil.removeNull(jarFiles).stream()
-                .map(File::getAbsolutePath)
-                .toArray(String[]::new);
-
-        if (GATEWAY_TYPE_MAP.get(SESSION).contains(runMode)) {
-            config.setJarFiles(jarPaths);
-        }
-
-        // 2.编译python
-        String[] pyPaths = UDFUtil.initPythonUDF(
-                udfList, runMode, config.getTaskId(), executor.getTableConfig().getConfiguration());
-
-        executor.initUDF(jarPaths);
-        if (ArrayUtil.isNotEmpty(pyPaths)) {
-            for (String pyPath : pyPaths) {
-                if (StrUtil.isNotBlank(pyPath)) {
-                    FlinkUdfPathContextHolder.addPyUdfPath(new File(pyPath));
-                }
-            }
-        }
-
-        Set<File> pyUdfFile = FlinkUdfPathContextHolder.getPyUdfFile();
-        executor.initPyUDF(
-                SystemConfiguration.getInstances().getPythonHome(),
-                pyUdfFile.stream().map(File::getAbsolutePath).toArray(String[]::new));
-        if (GATEWAY_TYPE_MAP.get(YARN).contains(runMode)) {
-            config.getGatewayConfig().setJarPaths(ArrayUtil.append(jarPaths, pyPaths));
-        }
-
-        try {
-            List<URL> jarList = CollUtil.newArrayList(URLUtils.getURLs(jarFiles));
-            // 3.写入udf所需文件
-            writeManifest(taskId, jarList);
-
-            addConfigurationClsAndJars(jarList, CollUtil.newArrayList(URLUtils.getURLs(otherPluginsFiles)));
-        } catch (Exception e) {
-            log.error("add configuration failed;reason:{}", LogUtil.getError(e));
-            throw new RuntimeException(e);
-        }
-
-        log.info(StrUtil.format("A total of {} UDF have been Init.", udfList.size() + pyUdfFile.size()));
-        log.info("Initializing Flink UDF...Finish");
-    }
-
-    private void writeManifest(Integer taskId, List<URL> jarPaths) {
-        FlinkUdfManifest flinkUdfManifest = new FlinkUdfManifest();
-        flinkUdfManifest.setJars(jarPaths);
-        flinkUdfManifest.setPythonFiles(FlinkUdfPathContextHolder.getPyUdfFile().stream()
-                .map(URLUtil::getURL)
-                .collect(Collectors.toList()));
-
-        FileUtil.writeUtf8String(
-                JSONUtil.toJsonStr(flinkUdfManifest),
-                PathConstant.getUdfPackagePath(taskId) + PathConstant.DEP_MANIFEST);
+        ExecutorContext.setExecutor(executor);
     }
 
     private boolean ready() {
@@ -309,46 +198,19 @@ public class JobManager {
         CustomTableEnvironmentContext.clear();
         RowLevelPermissionsContext.clear();
         FlinkUdfPathContextHolder.clear();
+        ExecutorContext.clear();
         return true;
     }
 
-    public ObjectNode getJarStreamGraphJson(String statement) throws Exception {
-        StreamGraph streamGraph = getJarStreamGraph(statement);
+    public ObjectNode getJarStreamGraphJson(String statement) {
+        StreamGraph streamGraph = JobJarStreamGraphBuilder.build(this).getJarStreamGraph(statement);
         return JsonUtils.parseObject(JsonPlanGenerator.generatePlan(streamGraph.getJobGraph()));
-    }
-
-    public StreamGraph getJarStreamGraph(String statement) throws Exception {
-        String currentSql = "";
-        DinkyClassLoaderUtil.initClassLoader(config);
-        String[] statements = SqlUtil.getStatements(statement, sqlSeparator);
-        ExecuteJarOperation executeJarOperation = null;
-        for (int i = 0; i < statements.length; i++) {
-            String sql = statements[i];
-            String sqlStatement = executor.pretreatStatement(sql);
-            if (ExecuteJarParseStrategy.INSTANCE.match(sqlStatement)) {
-                currentSql = sqlStatement;
-                executeJarOperation = new ExecuteJarOperation(sqlStatement);
-                break;
-            }
-            SqlType operationType = Operations.getOperationType(sqlStatement);
-            if (operationType.equals(SqlType.ADD)) {
-                AddJarSqlParseStrategy.getAllFilePath(sqlStatement).forEach(executor::addJar);
-                if (runMode.isApplicationMode()) {
-                    AddJarSqlParseStrategy.getAllFilePath(sqlStatement)
-                            .forEach(FlinkUdfPathContextHolder::addOtherPlugins);
-                } else {
-                    AddJarSqlParseStrategy.getAllFilePath(sqlStatement).forEach(executor::addJar);
-                }
-            }
-        }
-        Assert.notNull(executeJarOperation, () -> new DinkyException("Not found execute jar operation."));
-        return executeJarOperation.explain(executor.getCustomTableEnvironment());
     }
 
     @ProcessStep(type = ProcessStepType.SUBMIT_EXECUTE)
     public JobResult executeJarSql(String statement) throws Exception {
         Job job = Job.init(runMode, config, executorConfig, executor, statement, useGateway);
-        StreamGraph streamGraph = getJarStreamGraph(statement);
+        StreamGraph streamGraph = JobJarStreamGraphBuilder.build(this).getJarStreamGraph(statement);
         try {
             if (!useGateway) {
                 executor.getStreamExecutionEnvironment().executeAsync(streamGraph);
@@ -369,7 +231,7 @@ public class JobManager {
                 job.setResult(InsertResult.success(gatewayResult.getId()));
                 job.setJobId(gatewayResult.getId());
                 job.setJids(gatewayResult.getJids());
-                job.setJobManagerAddress(formatAddress(gatewayResult.getWebURL()));
+                job.setJobManagerAddress(URLUtils.formatAddress(gatewayResult.getWebURL()));
 
                 if (gatewayResult.isSuccess()) {
                     job.setStatus(Job.JobStatus.SUCCESS);
@@ -382,7 +244,8 @@ public class JobManager {
                 }
             }
         } catch (Exception e) {
-            String error = LogUtil.getError("Exception in executing FlinkJarSQL:\n" + addLineNumber(statement), e);
+            String error =
+                    LogUtil.getError("Exception in executing FlinkJarSQL:\n" + SqlUtil.addLineNumber(statement), e);
             job.setEndTime(LocalDateTime.now());
             job.setStatus(Job.JobStatus.FAILED);
             job.setError(error);
@@ -398,213 +261,20 @@ public class JobManager {
     public JobResult executeSql(String statement) throws Exception {
         Job job = Job.init(runMode, config, executorConfig, executor, statement, useGateway);
         ready();
-        String currentSql = "";
+
         DinkyClassLoaderUtil.initClassLoader(config);
-        JobParam jobParam = Explainer.build(executor, useStatementSet, sqlSeparator)
+        jobParam = Explainer.build(executor, useStatementSet, sqlSeparator)
                 .pretreatStatements(SqlUtil.getStatements(statement, sqlSeparator));
         try {
-            initUDF(jobParam.getUdfList(), runMode, config.getTaskId());
-
-            for (StatementParam item : jobParam.getDdl()) {
-                currentSql = item.getValue();
-                executor.executeSql(item.getValue());
-            }
-            if (!jobParam.getTrans().isEmpty()) {
-                // Use statement set or gateway only submit inserts.
-                if (useStatementSet && useGateway) {
-                    List<String> inserts = new ArrayList<>();
-                    for (StatementParam item : jobParam.getTrans()) {
-                        inserts.add(item.getValue());
-                    }
-
-                    // Use statement set need to merge all insert sql into a sql.
-                    currentSql = String.join(sqlSeparator, inserts);
-                    GatewayResult gatewayResult = submitByGateway(inserts);
-                    // Use statement set only has one jid.
-                    job.setResult(InsertResult.success(gatewayResult.getId()));
-                    job.setJobId(gatewayResult.getId());
-                    job.setJids(gatewayResult.getJids());
-                    job.setJobManagerAddress(formatAddress(gatewayResult.getWebURL()));
-                    if (gatewayResult.isSuccess()) {
-                        job.setStatus(Job.JobStatus.SUCCESS);
-                    } else {
-                        job.setStatus(Job.JobStatus.FAILED);
-                        job.setError(gatewayResult.getError());
-                    }
-                } else if (useStatementSet && !useGateway) {
-                    List<String> inserts = new ArrayList<>();
-                    for (StatementParam item : jobParam.getTrans()) {
-                        if (item.getType().equals(SqlType.INSERT)) {
-                            inserts.add(item.getValue());
-                        } else if (item.getType().equals(SqlType.CTAS)) {
-                            executor.getCustomTableEnvironment()
-                                    .getParser()
-                                    .parse(item.getValue())
-                                    .forEach(x -> {
-                                        executor.getCustomTableEnvironment().executeCTAS(x);
-                                    });
-                        }
-                    }
-                    if (!inserts.isEmpty()) {
-                        currentSql = String.join(sqlSeparator, inserts);
-                        // Remote mode can get the table result.
-                        TableResult tableResult = executor.executeStatementSet(inserts);
-                        if (tableResult.getJobClient().isPresent()) {
-                            job.setJobId(
-                                    tableResult.getJobClient().get().getJobID().toHexString());
-                            job.setJids(new ArrayList<String>() {
-
-                                {
-                                    add(job.getJobId());
-                                }
-                            });
-                        }
-                        if (config.isUseResult()) {
-                            // Build insert result.
-                            IResult result = ResultBuilder.build(
-                                            SqlType.INSERT,
-                                            job.getId().toString(),
-                                            config.getMaxRowNum(),
-                                            config.isUseChangeLog(),
-                                            config.isUseAutoCancel(),
-                                            executor.getTimeZone())
-                                    .getResult(tableResult);
-                            job.setResult(result);
-                        }
-                    }
-                } else if (!useStatementSet && useGateway) {
-                    List<String> inserts = new ArrayList<>();
-                    for (StatementParam item : jobParam.getTrans()) {
-                        inserts.add(item.getValue());
-                        // Only can submit the first of insert sql, when not use statement set.
-                        break;
-                    }
-                    currentSql = String.join(sqlSeparator, inserts);
-                    GatewayResult gatewayResult = submitByGateway(inserts);
-                    job.setResult(InsertResult.success(gatewayResult.getId()));
-                    job.setJobId(gatewayResult.getId());
-                    job.setJids(gatewayResult.getJids());
-                    job.setJobManagerAddress(formatAddress(gatewayResult.getWebURL()));
-                    if (gatewayResult.isSuccess()) {
-                        job.setStatus(Job.JobStatus.SUCCESS);
-                    } else {
-                        job.setStatus(Job.JobStatus.FAILED);
-                        job.setError(gatewayResult.getError());
-                    }
-                } else {
-                    for (StatementParam item : jobParam.getTrans()) {
-                        currentSql = item.getValue();
-                        FlinkInterceptorResult flinkInterceptorResult =
-                                FlinkInterceptor.build(executor, item.getValue());
-                        if (Asserts.isNotNull(flinkInterceptorResult.getTableResult())) {
-                            if (config.isUseResult()) {
-                                IResult result = ResultBuilder.build(
-                                                item.getType(),
-                                                job.getId().toString(),
-                                                config.getMaxRowNum(),
-                                                config.isUseChangeLog(),
-                                                config.isUseAutoCancel(),
-                                                executor.getTimeZone())
-                                        .getResult(flinkInterceptorResult.getTableResult());
-                                job.setResult(result);
-                            }
-                        } else {
-                            if (!flinkInterceptorResult.isNoExecute()) {
-                                TableResult tableResult = executor.executeSql(item.getValue());
-                                if (tableResult.getJobClient().isPresent()) {
-                                    job.setJobId(tableResult
-                                            .getJobClient()
-                                            .get()
-                                            .getJobID()
-                                            .toHexString());
-                                    job.setJids(new ArrayList<String>() {
-
-                                        {
-                                            add(job.getJobId());
-                                        }
-                                    });
-                                }
-                                if (config.isUseResult()) {
-                                    IResult result = ResultBuilder.build(
-                                                    item.getType(),
-                                                    job.getId().toString(),
-                                                    config.getMaxRowNum(),
-                                                    config.isUseChangeLog(),
-                                                    config.isUseAutoCancel(),
-                                                    executor.getTimeZone())
-                                            .getResult(tableResult);
-                                    job.setResult(result);
-                                }
-                            }
-                        }
-                        // Only can submit the first of insert sql, when not use statement set.
-                        break;
-                    }
-                }
-            }
-            if (!jobParam.getExecute().isEmpty()) {
-                if (useGateway) {
-                    for (StatementParam item : jobParam.getExecute()) {
-                        executor.executeSql(item.getValue());
-                        if (!useStatementSet) {
-                            break;
-                        }
-                    }
-                    GatewayResult gatewayResult = null;
-                    config.addGatewayConfig(executor.getSetConfig());
-                    if (runMode.isApplicationMode()) {
-                        gatewayResult = Gateway.build(config.getGatewayConfig()).submitJar();
-                    } else {
-                        StreamGraph streamGraph = executor.getStreamGraph();
-                        streamGraph.setJobName(config.getJobName());
-                        JobGraph jobGraph = streamGraph.getJobGraph();
-                        if (Asserts.isNotNullString(config.getSavePointPath())) {
-                            jobGraph.setSavepointRestoreSettings(
-                                    SavepointRestoreSettings.forPath(config.getSavePointPath(), true));
-                        }
-                        gatewayResult = Gateway.build(config.getGatewayConfig()).submitJobGraph(jobGraph);
-                    }
-                    job.setResult(InsertResult.success(gatewayResult.getId()));
-                    job.setJobId(gatewayResult.getId());
-                    job.setJids(gatewayResult.getJids());
-                    job.setJobManagerAddress(formatAddress(gatewayResult.getWebURL()));
-
-                    if (gatewayResult.isSuccess()) {
-                        job.setStatus(Job.JobStatus.SUCCESS);
-                    } else {
-                        job.setStatus(Job.JobStatus.FAILED);
-                        job.setError(gatewayResult.getError());
-                    }
-                } else {
-                    for (StatementParam item : jobParam.getExecute()) {
-                        executor.executeSql(item.getValue());
-                        if (!useStatementSet) {
-                            break;
-                        }
-                    }
-                    JobClient jobClient = executor.executeAsync(config.getJobName());
-                    if (Asserts.isNotNull(jobClient)) {
-                        job.setJobId(jobClient.getJobID().toHexString());
-                        job.setJids(new ArrayList<String>() {
-
-                            {
-                                add(job.getJobId());
-                            }
-                        });
-                    }
-                    if (config.isUseResult()) {
-                        IResult result = ResultBuilder.build(
-                                        SqlType.EXECUTE,
-                                        job.getId().toString(),
-                                        config.getMaxRowNum(),
-                                        config.isUseChangeLog(),
-                                        config.isUseAutoCancel(),
-                                        executor.getTimeZone())
-                                .getResult(null);
-                        job.setResult(result);
-                    }
-                }
-            }
+            // step 1: init udf
+            JobUDFBuilder.build(this).run();
+            // step 2: execute ddl
+            JobDDLBuilder.build(this).run();
+            // step 3: execute insert/select/show/desc/CTAS...
+            JobTransBuilder.build(this).run();
+            // step 4: execute custom data stream task
+            JobExecuteBuilder.build(this).run();
+            // finished
             job.setEndTime(LocalDateTime.now());
             if (job.isFailed()) {
                 failed();
@@ -614,7 +284,7 @@ public class JobManager {
             }
         } catch (Exception e) {
             String error = StrFormatter.format(
-                    "Exception in executing FlinkSQL:\n{}\n{}", addLineNumber(currentSql), e.getMessage());
+                    "Exception in executing FlinkSQL:\n{}\n{}", SqlUtil.addLineNumber(currentSql), e.getMessage());
             job.setEndTime(LocalDateTime.now());
             job.setStatus(Job.JobStatus.FAILED);
             job.setError(error);
@@ -624,47 +294,6 @@ public class JobManager {
             close();
         }
         return job.getJobResult();
-    }
-
-    public String addLineNumber(String input) {
-        String[] lines = input.split("\n");
-        StringBuilder sb = new StringBuilder();
-        for (int i = 0; i < lines.length; i++) {
-            sb.append(String.format("%-4d", i + 1));
-            sb.append("  ");
-            sb.append(lines[i]);
-            sb.append("\n");
-        }
-        return sb.toString();
-    }
-
-    private GatewayResult submitByGateway(List<String> inserts) {
-        GatewayResult gatewayResult = null;
-
-        // Use gateway need to build gateway config, include flink configeration.
-        config.addGatewayConfig(executor.getSetConfig());
-
-        if (runMode.isApplicationMode()) {
-            // Application mode need to submit dinky-app.jar that in the hdfs or image.
-            gatewayResult = Gateway.build(config.getGatewayConfig()).submitJar();
-        } else {
-            JobGraph jobGraph = executor.getJobGraphFromInserts(inserts);
-            // Perjob mode need to set savepoint restore path, when recovery from savepoint.
-            if (Asserts.isNotNullString(config.getSavePointPath())) {
-                jobGraph.setSavepointRestoreSettings(SavepointRestoreSettings.forPath(config.getSavePointPath(), true));
-            }
-            // Perjob mode need to submit job graph.
-            gatewayResult = Gateway.build(config.getGatewayConfig()).submitJobGraph(jobGraph);
-        }
-        return gatewayResult;
-    }
-
-    private String formatAddress(String webURL) {
-        if (Asserts.isNotNullString(webURL)) {
-            return webURL.replaceAll("http://", "");
-        } else {
-            return "";
-        }
     }
 
     public IResult executeDDL(String statement) {
@@ -761,6 +390,7 @@ public class JobManager {
         return Gateway.build(gatewayConfig).deployCluster();
     }
 
+    @Deprecated
     public JobResult executeJar() {
         // TODO 改为ProcessStep注释
         Job job = Job.init(runMode, config, executorConfig, executor, null, useGateway);
@@ -771,7 +401,7 @@ public class JobManager {
             job.setResult(InsertResult.success(gatewayResult.getId()));
             job.setJobId(gatewayResult.getId());
             job.setJids(gatewayResult.getJids());
-            job.setJobManagerAddress(formatAddress(gatewayResult.getWebURL()));
+            job.setJobManagerAddress(URLUtils.formatAddress(gatewayResult.getWebURL()));
             job.setEndTime(LocalDateTime.now());
 
             if (gatewayResult.isSuccess()) {
@@ -861,9 +491,5 @@ public class JobManager {
         }
         sb.append(statement);
         return sb.toString();
-    }
-
-    public Executor getExecutor() {
-        return executor;
     }
 }
