+++ dinky-core/src/main/java/org/dinky/job/builder/JobJarStreamGraphBuilder.java
@@ -1,0 +1,80 @@
+/*
+ *
+ *  Licensed to the Apache Software Foundation (ASF) under one or more
+ *  contributor license agreements.  See the NOTICE file distributed with
+ *  this work for additional information regarding copyright ownership.
+ *  The ASF licenses this file to You under the Apache License, Version 2.0
+ *  (the "License"); you may not use this file except in compliance with
+ *  the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ */
+
+package org.dinky.job.builder;
+
+import org.dinky.context.FlinkUdfPathContextHolder;
+import org.dinky.data.exception.DinkyException;
+import org.dinky.job.JobBuilder;
+import org.dinky.job.JobManager;
+import org.dinky.parser.SqlType;
+import org.dinky.trans.Operations;
+import org.dinky.trans.dml.ExecuteJarOperation;
+import org.dinky.trans.parse.AddJarSqlParseStrategy;
+import org.dinky.trans.parse.ExecuteJarParseStrategy;
+import org.dinky.utils.DinkyClassLoaderUtil;
+import org.dinky.utils.SqlUtil;
+
+import org.apache.flink.streaming.api.graph.StreamGraph;
+
+import cn.hutool.core.lang.Assert;
+
+/**
+ * JobJarStreamGraphBuilder
+ *
+ */
+public class JobJarStreamGraphBuilder extends JobBuilder {
+
+    public JobJarStreamGraphBuilder(JobManager jobManager) {
+        super(jobManager);
+    }
+
+    public static JobJarStreamGraphBuilder build(JobManager jobManager) {
+        return new JobJarStreamGraphBuilder(jobManager);
+    }
+
+    @Override
+    public void run() throws Exception {}
+
+    public StreamGraph getJarStreamGraph(String statement) {
+        DinkyClassLoaderUtil.initClassLoader(config);
+        String[] statements = SqlUtil.getStatements(statement, sqlSeparator);
+        ExecuteJarOperation executeJarOperation = null;
+        for (int i = 0; i < statements.length; i++) {
+            String sql = statements[i];
+            String sqlStatement = executor.pretreatStatement(sql);
+            if (ExecuteJarParseStrategy.INSTANCE.match(sqlStatement)) {
+                executeJarOperation = new ExecuteJarOperation(sqlStatement);
+                break;
+            }
+            SqlType operationType = Operations.getOperationType(sqlStatement);
+            if (operationType.equals(SqlType.ADD)) {
+                AddJarSqlParseStrategy.getAllFilePath(sqlStatement).forEach(executor::addJar);
+                if (runMode.isApplicationMode()) {
+                    AddJarSqlParseStrategy.getAllFilePath(sqlStatement)
+                            .forEach(FlinkUdfPathContextHolder::addOtherPlugins);
+                } else {
+                    AddJarSqlParseStrategy.getAllFilePath(sqlStatement).forEach(executor::addJar);
+                }
+            }
+        }
+        Assert.notNull(executeJarOperation, () -> new DinkyException("Not found execute jar operation."));
+        return executeJarOperation.explain(executor.getCustomTableEnvironment());
+    }
+}
